requirements.txt
requests
python-dotenv
regex
pytest
